gui/loggingWindow: remove "empty space" between STDOUT and STDERR panes.
The empty space came from having a top and bottom pane, but no center pane.
Fixed by setting one of them to be a CenterPane. Unfortunately, it seems that
MinSize is not respected for a CenterPane but this is still better than
having to resize both panes, when we only want to resize one.