Added polarizerHandler to relevant experiments.
Previously, we were using the angle handler, but the usage of
angleHandler in structuredIllumination is not wavelength-aware.
I considered the following possibilities deal with multiple
wavelengths.
a) Moving angleHandler stuff into expose() where actions can be
   defined with a wavelength. This would break any
   existing angleHandler that doesn't expect a wavelength.
b) Creating a top-level angleHandler that can take actions
   both with and without wavelength (perhaps tracking wavelength,
   as well), and delegate them to lower-level angleHandlers.
   This would add significant complexity, and I'm not sure the
   benefit is worth it.
c) Changing the polarizer handler from 'SI angle' to a new
   'SI polarizer', and adding it to experiments where required.
   This requires replacing angleHandler with polarizationHandler
   in rotatorSweep, and adding new angleHandler code to
   structuredIllumination, and seems like the simplest approach.