util/Mrc.py: always shape all data if file truncated (bug #289)
In case of truncated files, find which dimension needs to be truncated
instead of only truncating the first dimension.  In addition, introduce
any padding required to shape all data in file.

For example, if data should had shape (5, 10, 512, 512) and file was
truncated, previous behaviour would shape data in (N, 10, 512, 512).
If there was only (1, 5, 512, 512) worth of data, then it would shape
(0, 10, 512, 512) and data would be empty.

Also, use ndarray.view() instead of assigning to __class__ for
ndarray_inMrcFile because the concatenating of padding data means that
data is now a numpy.array and not a numpy.memmap and class assignment
is only for heap types.