maint: assign copyright holders to each file (issue #308)
UCSF code is under 3-clause BSD license but our changes at Micron are
under GPLv3+.  Go through the git history of the project and place the
BSD license header on the files we got from UCSF.  Then assign
copyright under GPLv3+ to who wrote the rest of the code.

The project is now under GPLv3+ so remove BSD license at the root of
the repository and leave it only on the files.