Move all cockpit application/library into a cockpit package.
In addition to moving all files into a subdirectory/package named
cockpit, the cockpit.pyw is now cockpit/__init__.py.  This is the
first step on preparing cockpit to be distributed with setuptools.
All of the imports will still need to be fixed so tha they are not
relative the current directory.