Change all imports to be based on the new cockpit package.
Keep the imports that were relative to module current directory the
same, i.e., leave the 'from . import x' but others needed to be
relative to package root, i.e., cockpit.  Also added new imports
(testing import of all modules brought up several missing ones) and
remove imports that were not actually required.

Also, fix the dynamic importing in cockpit/util/importer.py
Now that this is a python package, we need to be able to import this from

Because cockpit is now a python module and not a pyw script, remove
the kludge on cockpit/testsuite/test_sim_experiment.py