depot.py: fix location of devices folder in DeviceDepot.initialize()
Use absolute path for cockpit.device sub package when listing all its
files.  This prevents cockpit from working now that it's installable
via setuptools and we probably won't be running from the root of the
package.  Also use glob which does the *py filter.

Remove the module variables THIS_FOLDER which was only used to be
joined with the DEVICE_FOLDER and rename it to FPATH because it now
has the absolute path.

Avoid checking for files logic when checking for device.configurator.
Just try to import it and catch the error for missing module.