dataSaver: write more metadata in extended header like softWoRx (issue #290)
We want to write dv files but we write the extended header metadata
in our own non-specified format (basically, only the timestamp).
This changes to write an extended header that is familiar to softWoRx
so we can then use it for analysis.  This only puts the right shape
of the extended header, we still need to figure out how to obtain the
rest of the metadata which currently is not avaliable to
DataSaver.writeImage().

In addition, move computations with imageData outside the file lock,
and reduce try/catch to the actual writing of the file.