ImmediateModeExperiment: reduce code duplication from Experiment
Large chunk of the ImmediateModeExperiment.run() was duplicated from
the parent Experiment.run().  This was because ImmediateMode does
not want to prep the device handlers or create action table which is
done in Experiment.run().  However, those are Experiment methods on
their own (prepareHandlers() and createValidActionTable()).  So
overload those methods in ImmediateModeExperiment with empty and
make use of the parent run().

The only expected change is that immediateMode will now perform
extra checks on camera readout times.  Maybe this should be moved
out of run() and into one of the methods run() calls.

Also, make use of dict comprehension, remove trailing whitespace,
and cut some lines to under 80 cols.