dataSaver: include excitation wavelength in extended header (issue #290)
This assumes that all images from a camera have the same excitation.
We can assume that, it's a limitation of the cockpit interface.

This also assumes that each image has only one excitation wavelength
worth of being included in the metadata.  We need to assume that,
that's a limitation the DV format, even if cockpit does allow multiple
light sources for each image.