Made callInMainThread re-entrant.
callInMainThread is a decorator that pushes function calls to the main
thread, to be called after any pending wx event handlers. If one
function decorated this way calls another, then the second call was
unnecessarily requeued. In certain cases, this could prevent the outer
call completing as expected.

I've added a test so that this re-entrant behaviour doesn't result in
requeuing if execution is already in the main thread. We can now add
this decorator to *any* function that touches the UI and therefore
needs to be in the main thread, rather than having to find the outermost
call that first touches the UI.