Remove canvas width and height, use GetClientSize (issue #368)
The width and height properties of MosaicCanvas and SlaveCanvas are
only set after initGL which is only set after the canvas receives a
paint event.  This causes issues during startup because zoomTo() may
be called before the paint event is handled (see issue #368).  This
replaces the width and height properties with wx calls which is
already tracking those sizes for us, even before the cnvas has been
painted.

Also, replace setSize with directly calling SetClientSize which also
triggers a paint event and so no need to explicitely call Refresh().