Fix overflow error by wrapping non-python types.
One or more operands in the calculation of dataOffset are int32 rather
than a python int. This causes an overflow error when the value of
dataOffset exceeded 2**31-1, meaning that data beyond 2Gb in any
experiment is lost. Wrapping these operands with int() solves this
problem.

planeBytes is wrapped once, on assignment.

I think headers[fileIndex].next has to be wrapped each time, as this
value can be reassigned by util.Mrc. We should perhaps fix this in Mrc.py.