Use getter for width and height.
Occasionally, width and height could be referenced before they were
intialised in MacroStageXY - they were not initialised in __init__, so
were only set after the first call to InitGL, which in turn calls
GLCanvas::GetClientSize. Just use this getter where it's needed - that
saves two member variables, and ensures that the width and height used in
calculations is current.