interfaces.stageMover.optimisedSiteOrder: always return a new list and not a ref
If the new distance is longer, we return the same order as the input.  If so,
make sure we return a copy of the list and not the same list.

Also, note that on the name, i.e., use `optimised` instead of `optimise`
(just like Python's `sorted` and `list.sort`).  Also spelling.