stageMover: copy handlers to have same number of handlers per axis (issue #413)
stageMover is not great when there are different number of handlers
for each axis (e.g., one XY axis and two Z axis).  Functions need to
check a missing mover or use the closest one.  The "right" fix might
be to have each axis being selected independently (instead of a single
index that selects all movers from the three axis lists).  In the mean
time, we can hack this by making copies of the movers.