cockpit/devices/myZMover.py.TESTING: remove file
This file provides a Faze Z mover which is what the dummyZStage does
and was added in a3aa8a34bb79 around the same time as dummyZStage
(which was added in f154c18e6527).  Its source is also quite similar
to dummyZStage.  Probably an earlier version of the it, and
accidentally added.  It was renamed TESTING in 50b5706231b0 probably
also accidentally.  Anyway, seems like it does nothing so we are
removing it.