Single place for default data/log/config directories (issue #320)
This also means that dataDirectory, logDirectory, and configDirectory
are no longer "required" keys for the Configurator, and are just
options that change the default.

cockpit.util.files already defined a default value for those things.
So they got removed from the default Configurator class.  Those values
in the configurator are not actually used again, everywhere else, it's
the cockpit.util.file functions that provide the interface.  So keep
we the defaults there.

Also, no one else is making using of the ROOT_DIR, DATA_DIR, LOGS_DIR,
and CONFIG_DIR values, so don't signal them as public.