cockpit.util.logger: don't use login name for log files.
Each user should have their own directory where cockpit will save the
log files, so there is no need include the username on the log file
and separate between the 'global' log file (before login) and the user
log file (after login). It is also not possible to change user, one
needs to exit and restart cockpit, so there would always be a separate
log file.

This whole thing only made sense when there was a single directory,
writtable by everyone, to store the log files, and user identification
was handled by cockpit instead of the OS.

This change also removes the generateLogFileName function and
curLogHandle global variable which are no longer used.