Remove user login and user logout events (issue #293)
There is no longer a cockpit to login and logout so this events no
longer make sense.  The login is used by devices to use the last
device and window settings.  Because userConfig is loaded even before
the wx App starts, this is no longer necessary.

The logout is used by some devices to close connections.  We replaced
it with 'program exit' event.