Move window positioning out of cockpit.util.user
Move the position of all the app windows into the App instance.  This
leaves cockpit.util.user almost empty.

Handle the camera view title when saving userConfig and not when
reading.

The main window was handled in a special manner because it was
supposed to already be destroyed by the time the program ends.
However, testing shows that it is no longer true so remove the whole
special case.

The mosaic window was also looking for Rect instead of Position, but
that was never saved to the userConfig.  So remove that special case
as well.