Remove cockpit.util.user module (issue #293)
Remove module which is not doing much anymore, it only had a logout
method which was called by App.OnExit.  Move its code elsewhere.

Move the writing of the loggingWindow text to the logger into the
loggingWindow.  The loggingWindow getStdErr and getStdOut functions
are no longer used so remove them.