cockpit.util.importer: remove module that imports all modules in package
This module is used to import and act all on all modules of a specific
subpackage.  It was historically useful to handle cockpit.devices.
However, cockpit.devices no longer works in that way so the importer
is currently only used by cockpit.interfaces.  But the interfaces are
not meant to be dynamic like the devices used to be, and there's no
need to do this by name.

So remove the imported module and initialize/makeInitialPublications
imager and stageMover by name.