cockpit.gui.camera.viewPanel: don't specify priority to 'filter change' event
The very high priority value would make the subscription by ViewPanel
run later.  A comment mentions this was required to ensure that the
ViewPanel responds after the cameras.  However, this is out of date.
The comment is talking about an older subscription to 'drawer change'
which was removed in 3750be06d816 .  Currently, nothing subscribes to
'filter change' so the priority value is pointless.