cockpit.events: rewrite into Publisher and OneShotPublisher classes (issue #461)
The cockpit.events module is like a singleton class using global
variables and free functions instead of class properties and methods.
This is more limited since we can't have more than one instance of an
events class and complicates test cases with multiple patches.  In
addition, there's a fair amount of code duplication between the normal
publisher and one shot publisher.

This commit rewrites the events module into the new Publisher and
OneShotPublisher classes keeping a singleton instance of each for
backwards compatibility.  It adjusts the patches in the test suite for
this module.  The only code left from before is the executeAndWaitFor
functions which were added by Mick Phillips so we remove UCSF
copyright and keep Mick's.