ExceptionBox: don't hide exception traceback in collapsible pane (issue #467)
In most cases, the error message is not enough, the full traceback is
needed to understand the error.  So we want to have the traceback
displayed by default.  If we have the collapsible pane open by
default, then we might as well have no collapsible pane at all.  This
change removes it.