cockpit.main(): show GUI with exc caught during initialisation (issue #439)
There are initialisation steps before CockpitApp.OnInit.  If an
exception occurs then cockpit exits with errors which the user does
not see unless they are on the command line.  At this stage, there is
no wx.App created to display a dialog with the exception.

This change creates a simple wx.App with the purpose of displaying the
exception traceback.