Remove duplicated TRIGGER_* "enums" from microscope.devices.
`cockpit.handlers.camera` defines a kind-of-enum for trigger types.
This appears duplicated on the `cockpit.devices.microscopeCamera` and
`cockpit.devices.softCamera` modules but those are already diverging
(on softCamera it's named TRIGGER_SOFTWARE instead of TRIGGER_SOFT and
both copies are missing TRIGGER_DURATION_PSEUDOGLOBAL).

This commit removes the copies of trigger types enum and replaces it's
small usage with the one from `cockpit.handlers.camera`.