statusLightsWindow: replace status lights window with status bar (issue #565)
This commit removes the status lights window and replaces it with
individual fields on a new status bars for both the main window and
touchscreen window.

All usage of status lights were using the yellow colour.  The only
exception is grey which was used to reset to an hardcoded default
background colour.  Since there is only one colour, this commit also
removes the backgroundColour argument for the UPDATE_STATUS_LIGHT
event.  The colour choice is now done only on the StatusBar itself.

In addition, because status bar fields cannot have individual
background colours, change of colour is applied to the whole bar.