Replace 'update status light' event name with corresponding "enum".
It used to be that events would be subscribed and published by
name, and each action would have to use the correct string.
Since 242f399c441e it was added cockpit.event variables with
those names, a sort of "enum".  However, actually replacing the
usage of the event names with those enums is an ongoing process.

This commit replaces all usage of the 'update status light' event
name with the corresponding cockpit.events "enum".