CamerasWindow: remove blank view unless there are no active camera (issue #545)
Enabling and disabling of cameras used to be done via the selector
button at the top of a blank camera view.  This required that there
was always one extra ViewPanel.  This is undesirable since it takes up
space on the screen (issue #567).  Currently, enabling and disabling
of cameras is done via the main window so the extra ViewPanel is no
longer required.

This commit removes the extra ViewPanel unless there are no enabled
cameras since otherwise the window would have no content.