events: replace 'user abort' with events.USER_ABORT (issue #570)
Events are identified by name (a string). Using events required
knowing the exact string and if there was some typo nothing would
happen since it was just assumed there was just no subscriber to that
typo-ed event name.  Since 242f399 there are global constants in the
cockpit.events module to prevent this.

This commit replaces use of the 'user abort' event name with the
events.USER_ABORT constant.