CockpitApp.onExit: fix UnboundLocalError for cockpit variable.
import of a cockpit module inside a function causes issues because it
means that the global cockpit symbol is replaced with a local one.
This means the cockpit subpackages in the global symbol might not be
available in the local cockpit symbol, and even that it is assigned
until after the local import leading to UnboundLocalError, local
variable 'cockpit' referenced before assignment.  See discussion of
issue #529.

This commit moves the import to the global scope and avoid this issue.
This has been happening for some time but only at exit so no one has
noticed.  It can be seen on the error log files and command line
because it was also leading to segfaults.