events: replace event names with events global constants (issue #570)
Events are identified by name (a string). Using events required
knowing the exact string and if there was some typo nothing would
happen since it was just assumed there was just no subscriber to that
typo-ed event name.  Since 242f399 there are global constants in the
cockpit.events module that prevent this and should be used instead.

This commit replaces all usage of event names for events to which
there is a corresponding constant (some events don't yet have one).