cockpit/resources/bitmaps/: remove unused icon files.
With the exception of cockpit.ico, the icons are for the touchscreen
UI but most are actually unused.  All the laser icons (filenames
laser_*.png) were replaced in 41d5f3aa4347 with icons generated on the
fly.  The trans_light.png, delete_tiles.png, and find_stage.png were
never actually used.

This commit removes the icon files that are not actually in use.