cockpit/devices/imager.py: replace with Microscope DummyDSP (issue #572)
cockpit/devices/imager.py only has the DummyImagerDevice class.  We
have a more complete DummyDSP class on Microscope which can be used to
actually fake experiments, so the aim is to replace DummyImagerDevice
with an instance of DummyDSP, like it was done for DummyCamera (see
d9932a15d114).

This commit renames the dummyCamera module into dummies with the plan
to have all the dummy device classes.  To avoid code duplication, most
of the previous DummyCamera is now _MicroscopeTestDevice class.  The
new DummyDSP class is also created to replace the DummyImagerDevice.