events: remove unused 'dummy take image' event.
With the removal of DummyImagerDevice (77942cdfbb93) and rewrite
of DummyCamera (d9932a15d114), there is nothing else publishing
the 'dummy take image' event.  The CameraHandler cared about it
in case one of its devices was a DummyCamera.  The SoftCamera
cared about because it was used in a device without a real
trigger source, but that was replaced with actually using the
softTrigger callback.

This commit removes subscribe/unsubscribe of 'dummy take event'
which is no longer being published and fixes it's questionable
use in softCamera.