MosaicWindow: set reasonable default size (issue #574)
The default size for MosaicWindow does not take into account its
contents.  The default size was previously hardcoded but that was
removed with 27e6d407978b.  Part of the issue is the difficulty of
automate a BestSize for a GLCanvas such as MosaicCanvas.

This commit sets the default Height for MosaicWindow based on the
buttons on the left hand side panel which happens automatically with
correct use of a sizer after removal of the parent Panel.  The Width
is set to 4 times the left hand side panel, so that the mosaic takes
3/4 of the width.