AmbientLight: new device class for the room or ambient light (issue #580)
The ambient light source is added directly to the list of handlers
during DeviceDepot.initialize but does not have an associate device.
The plan is to not have the ambient light and require users to specify
it on the depot.conf file which requires an associated device class.

This commit adds an AmbientLight device class and instantiates it just
like other devices on DeviceDepot.