depot: do not automatically always add an ambient light device (issue #580)
We don't want to always have an ambient light, it should be a
configuration.  Also, because there are now DummyLasers added in the
absence of other light source devices, AmbientLight is no longer
necessary as fallback either.

This commit removes the AmbientLight device from being automatically
added to devices and requires users to specify it on depot.conf files.