MainWindow: don't manually set background colour to the default.
For some reason, SYS_COLOUR_WINDOW (window background colour) in GTK3
is white which is incorrect which leads to white background on the top
two rows of the main window and grey on the bottom row with the light
sources.  Anyway, the best way to set a panel background colour to the
window default, is to do nothing and have it inherit from the parent.

This commit simply removes the setting of background colour.