MainWindow: fix size issue under GTK3 which hides light source SetPointGauge
wxWidgets has size issues under GTK3 if the default font size is
modified (see https://trac.wxwidgets.org/ticket/16088).  We use a
larger font on the main window panels and the final effect is that the
SetPointGauge at the bottom of the light source panel is hidden.

This commit uses the upstream recommended workaround by redoing the
layout and sizing from a wxShowEvent handler.