Save window state (shown or hidden) only when cockpit exits (issue #586)
The function that saves the window state only does it when the window
state is changed from the martial windows menu and not when it is
closed by the close box on the frame.  The exception is the shell
window which also does it when it's shown/hidden via the keyboard
shortcut.

This commit changes it so it is only saved when cockpit exits.