Only use bool for windowState userConfig.
The userConfig windowState entry, which controls whether a window is
shown or hidden during cockpit startup, is mostly a 0/1 value.  Except
when the python shell window is shown or hidden via the keyboard
shortcut (`keyboard.showHideShell`) in which case it is saved as bool.

This commit ensures that the config value is always set as bool.