CockpitApp.OnInit: fix decision whether to show or hide window (issue #586)
At the moment, the CockpitApp expects all windows to show themselves,
and then only calls Hide if their window state user config is False.
This means that if a window does not call Show on self, even if window
state is True, the window won't be shown.  This is the case of the
touchscreen and intensity tool.

This commit changes the logic so that cockpit will call Show or Hide
appropriately and the windows themselves don't have to worry about it.