ViewPanel: remove setImageSize from camera view selector (issue #569)
The selector button on the cameras ViewPanel has the ability of
showing possible image sizes and setting them on the camera.  This
requires having getImageSizes (note the plural) and setImageSize on
the device class.  These work by mapping a string, usually something
like "512x512" to the actual size.  These are not implemented in our
MicroscopeCamera device class.

This commit removes the feature since it's not implemented for some
time and no one has noticed.