mainWindow: change button font via the button font instead of system default.
The buttons on the top row of the main window are larger than default.
This is done by getting the default font (via wx.SystemProperties) and
modifying it.  This means that any changes to the font from the button
parent is lost.

This commits sets the font of the main window buttons to a larger
version of its own font, not to a larger version of the default system
font.