Imager.videoMode: add timeout when waiting for a new image (issue #584)
If a camera drops a frame, the video threads gets blocked waiting for
a NEW_IMAGE event.  The existing fix of clearing one shot subscribers
does not work since it is only called after unblocking the waiting for
the event.

This commit fixes the issue by adding a timeout to the wait.