cockpit.gui.devices: use relative instead if hardcoded size fonts.
Using hardcoded size fonts when we want default and smaller than
default fonts means that UI won't respect the size requested by the
user and ignore the font size ignored by the parent.  Instead,
manually get the font and modify its size and attributes.

In addition, instantiating a wx.Font requires that a wx.App be created
first.  A side effect of this change is that this module can now be
imported before.