Remove separation between primary and secondary windows (issue #593)
Whether a window is primary or secondary changes whether a window can
be hidden at all, the idea being that a primary window is so important
that a user should not be able to hide it.  However, turns out that
different people have different opinions about which set of windows
are the most important.

This commit removes the whole primary vs secondary windows and all
windows are treated the same, with the exception of the main window.
Whether a window should be shown or hidden at start is now defined on
the user configuration with an optional Frame SHOW_DEFAULT setting a
default.  This commit also changes the martialWindows menu so all
windows all have the same submenu.