MainWindow: split into MainWindowPanel to fix background on Windows (issue #592)
On Windows, upstream recommends against naked Frames, i.e., against
placing the Frame content into a sizer that is set on the Frame
directly.  Instead, we are supposed to have a Panel with all that
content and that sizer, and have that Panel as the Frame single child.
Upstream says it prevents varies issues, but in our case it solves the
issue of a dark grey bar between the top panel and the lights panel.
This appears because the default background colours of a Frame and
Panel are dark and light grey respectively, so the space between the
two panels show the dark grey from the Frame.

This commit splits the code with the MainWindow content into a new
MainWindowPanel, and keeping the MainWindow class for the Frame
specific code.