LoggingWindow.WriteToLogger: remove writing of LoggingWindow when Cockpit exits.
When Cockpit exits it writes the content in the LoggingWindow to the
log file.  However, text on the LoggingWindow is already written to
cockpit.util.logger.log while cockpit runs and so is already on the
log file when cockpit exits.

This commit removes the call to LoggingWindow.WriteToLogger from
CockpitApp.onExit which led to duplicated content on the log files.
It also removes the LoggingWindow.WriteToLogger method which is no
longer used.