mainWindow: move "light path" button to the main window menu bar (issue #597)
This commit moves the "light path" logic, now renamed to "channels",
from an OptionButtons on the main window panel to a menu on the main
window menu bar.  It also changes the available options: instead of
saving a single channel to file, one now exports all of the channels
to a file; the same for load/import; instead of updating an existing
channel, we just save with the same name.