MacroStageWindow: use wx controls to draw positions and step size (issue #594)
Current step size is drawn with OpenGL and FTGL.  Same with the
position of each stage.

This commit replaces that with wxTextCtrl.  Using wx controls will
make it look native and easier to interact in the future.  It also
reduces the use of FTGL and OpenGL.  The wx controls are larger which
makes the multiple window elements to be misaligned.  A separate
commit should reorganize the window content.