Move primitives configuration from stage device to cockpit.conf (issue #605)
The primitives configuration is about what to draw on top of the
stage, things such as the shape of wells in a plate.  This was defined
in the stage device configuration but makes little sense now that we
have a cockpit configuration separate from the devices.

This commit removes the primitives configuration from all stage
devices and associated methods from PositionerHandler, and stageMover.
This commit also adds a primitives config to the stage section of
Cockpit configuration.