StageMover: add SavedTop and SavedBottom properties (issue #421)
The saved top and bottom positions are stored as variables in the
cockpit.gui.saveTopBottomPanel module.  Other GUI components need this
data which is odd: GUI components should not have to access other GUI
components to access data, GUI components should only be a view on
that data.

This commit moves the storage of the currently saved top and bottom
position to the StageMover class.