moveZCheckMoverLimits: move from savedTopBottomPanel to stageMover.
The function moveZCheckMoverLimits has nothing to do with the
savedTopBottomPanel, it's not even part of the GUI.  Its location on
the savedTopBottomPanel module is only because historically was only
used by that panel.  It is now also used by the touchscreen.

This commit moves the moveZCheckMoverLimits function to the StageMover
interface which handles that type of logic.