New STAGE_TOP_BOTTOM event when updating saved top/bottom position (issue #421)
When the saved top and bottom positions are changed, this needs to be
updated on all MacroStageZ (macro stage and touchscreen windows) and
on the SavedTopBottomPanel text controls.

This commit has StageMover publishing an event so that all windows
that display the saved top and bottom positions in any form can update
themselves.