Use super().method(...) to call inherited methods.
For Python 2 compatibility we use super(cls, self).method(...), and
for old style class compatibility we use parent_cls.method(self, ...).
However, we now require Python 3 so we no longer need to keep that
compatibility.  That old usage is not only more verbose, it leads to
issues when we change the class hierarchy and forget to update the
call.

This commit replaces them all with super().method(...).