Remove check for new primitives on each stage paint (issue #608)
There is code in place that each time a stage is drawn checks for new
primitives to draw.  This is a new feature that is not finished yet
(see issue #608) and so the checks are not needed.

This commit removes the few lines of code related to it until that new
feature is implemented.