StageMover.onMotion: remove unused arguments and update STAGE_MOVER events.
StageMover.onMotion is the only subscriber of STAGE_MOVER events and
does not make use of the deviceName and position arguments.  The first
is completely not required and the later is actually queried from the
handler again.

This commit removes the deviceName and position arguments from
StageMover.onMotion.  It also updates all publications of STAGE_MOVER
events to not include them.