PositionerHandler: remove required setSafety callback (issue #611)
The setSafety call is used to set software limits on the hardware if
they support it.  But cockpit itself also checks the soft limits and
will error first so moves beyond the soft limits shouldn't reach the
device's software.  In addition, only the Nanomover even bothers to
actually do anything on setSafety.

This commit removes the requirement for the setSafety callback as well
as the associated methods (mostly empty) on the stage devices.