CockpitApp.OnInit: ensure main window is shown when program starts.
The Cockpit main window does not hide like the others (it might be
minimized/iconized but that is a separate thing).  As such, we
shouldn't be saving its shown state and shouldn't restore it at start
like we do for the other windows.  If we do, we risk having a
situation where the user has somehow hidden the main window which
means it won't appear next Cockpit starts.

This commit limits the whole show/hide logic to windows other than the
main window.