Channels: new interface/container class for channel configurations (issue #597)
The container for the different channel configuration is within
ChannelsMenu itself.  This coupling of the GUI (menu) and the
container makes it difficult to have another view of the same
container such as the planned column of buttons to quickly switch
between configurations.

This commit moves the container itself from the ChannelsMenu into a
new Channels class in the interfaces sub-package.  This new class is
effectively a container but it is also a wx.EvtHandler that will emit
wx events when its content changes.  Views of the container can bind
to those events and update themselves.