Experiment: remove cameras and lights arguments from constructor (issue #409)
The constructor for Experiment takes a list of handlers for cameras
and lights.  It also takes exposureSettings which maps those handlers
to exposure times.  Because the required camera and light handlers are
already in exposureSettings they don't need to be passed as separate
arguments.

This commit removes the cameras and lights arguments from the
Experiment constructor and adjusts the ExperimentConfigPanel
accordingly.  In addition, it changes ImmediateModeExperiment which
for some reason passes an empty exposureSettings.