LoggingWindow: reparent to the main window after that is created.
The logging window is created before the final main window so that it
can show any output during initialisation.  This means it's created
before the main window and so, unlike all other Cockpit windows, it is
not a child of the main window.

This commit reparents the logging window to the main window after it
is created.  Among other things it ensures that the logging window is
raised to top like all others during CockpitApp.onActivateApp.