WindowsMenu: move menu to manage windows to the main window menu bar(issue #583)
There is a menu to manage the position and state of the multiple
cockpit windows.  This menu is accessed via right click on the main
window background or via Ctrl+P, none of which provide good
discoverability.  The Cockpit main window now has a menu bar so this
menu could be moved there do address this.

This commit removes cockpit.gui.keyboard.martialWindows, the function
that creates such menu, and creates a WindowsMenu class for the main
window menu bar.  The shortcut also changes from Ctr+P to Alt+W.  With
this commit the menu is created only once while previously a new menu
was created each time.  Because most windows do not exist when the
menu is created, the new menu updates itself each time it is opened.