Channels: use ProcessEvent instead of QueueEvent for channel events.
Using QueueEvent instead of ProcessEvent means the events are handled
asyncronous.  This creates problems if the Channels instance is used
before the events are processed (see example on the documentation
added on this commit).

This commit uses ProcessEvent instad of QueueEvent for the
EVT_CHANNELS_ADDED and EVT_CHANNELS_REMOVED events.