SaveTopBottomPanel: remove hardcoded button sizes and add space between widgets.
The buttons and text controls in the SaveTopBottomPanel have an
hardcoded size is which is barely enough for default Windows and Mac,
and not enough at all on default Gnome.

This commit replaces the hardcoded sizes with a sizer and expand flags
to ensure they all have the same size and things are aligned.  It also
adds the default wx borders.