SaveTopBottomPanel: use disabled TextCtrl to show height.
The height of the currently saved top and bottom are shown with a
StaticText but that is better suited for content that does not change
over time such as labels.  A disabled TextCtrl is a more common
element for something that may change but the user can't set directly.

This commit replaces the StaticText with a disabled TextCtrl.