ChannelsPanel: manage space from TopLevelParent when adding/removing button.
If adding/removing a button requires changing the number of columns,
we need to call Layout() on the TopLevelParent for its sizer to set
the space.  This is specially noticeable when adding new buttons since
otherwise the new column of buttons is invisible until the frame is
manually resized.  Similarly, we might need to make the window wider.