MacroStageWindow: re-organise window (issue #594)
This commit joins the xy and z position controls in a 3 row panel, and
aligns it with the save top bottom panel; removes the static box
around the save top bottom panel; uses default panel background colour
instead of forcing white (this means that the active TextCtrl which
have white background will "attract" the eye while the Disabled fade
on the background instead of the other way around); and places the
bottom row of buttons aligned under the macro stages themselves.  The
new layout doesn't use GridBagSizer so it also removes comments
apologising for its use.