Replace list of step sizes with guess best next/previous step size (issue #612)
Each device creates its own list of step sizes and default step size
which is then handled by the PositionerHandler.  The list is arbitrary
and depends on what the developer of the device class felt was best at
the time.  Besides the fact that the user may feel different about
what step size values are sensible, this is a problem for a device
class that interfaces with multiple stages of different sizes, such as
the planned Python Microscope stages (see issue #614), as it would
require a way to guess the best step sizes based on stage size or
extra configuration.

Moving by a step is a higher level from relative moves that does not
exist on the devices classes and is implemented on the handlers only.
But handlers are the level that abstracts the difference between
different devices and if steps don't exist on devices they shouldn't
exist on the handlers either.  Instead, the moving by a step operation
should be at the cockpit interfaces level.

This commits removes the whole step size logic from the stage devices
and positioner handlers and implements it on the StageMover interface.
Instead of a list of step sizes, the StageMover only keeps track of
the current sizes and has logic to estimate a sensible increase or
decrease of the step size.  This also removes the limits on the step
size --- too large and too small steps will be handled by the move
operations.