MacroStageBase: don't redraw when step size changes, only redraw MacroStageZ.
The macro stages used to display the step size value but that is
now its own wx control.  The "crosshair" on the MacroStageZ is
the range of the step size so only MacroStageZ needs to redraw
when it changes.

This commit moves the subscription to 'stage step size' from
MacroStageBase to MacroStageZ and slightly changes the logic on
MacroStageBase.refreshWaiter to accommodate it.